### 2025-07-30 #8668 @crvernon
:wave: @yk748 - your paper needs to be in markdown, not Rmarkdown and have an accompanying paper.bib file.  Please review the submission guidelines to ensure this is correct and send me a comment here when this has been corrected.
