+++ paper.Rmd
@@ -1,74 +1,0 @@
----
-title: 'cxreg: An R package for complex-valued lasso and graphical lasso'
-tags:
-  - Spectral analysis,
-  - Regularization path,
-  - Coordinate descent ,
-  - lasso, 
-  - Graphical model
-authors:
-  - name: Younghoon Kim
-    orcid: 0009-0007-0117-5530
-    equal-contrib: true
-    affiliation: 1
-  - name: Navonil Deb
-    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
-    affiliation: 1
-  - name: Sumanta Basu
-    corresponding: true # (This is how to denote the corresponding author)
-    affiliation: 1
-affiliations:
- - name: Department of Statistics and Data Science, Cornell University, United States
-   index: 1
-date: 19 July 2025
-bibliography: cxreg-ref.bib
----
-
-# Summary
-
-*Pathwise coordinate descent* [@friedman2007pathwise] is a popular iterative optimization method for solving penalized linear regression [*lasso*; @tibshirani1996regression] and penalized Gaussian log-likelihood [*graphical lasso*; e.g., @banerjee2008model] problems. Given a specified or derived sequence of tuning parameters $\lambda$ in the penalty term, the algorithm updates one coordinate at a time by isolating the target variable using partial residuals.
-
-The coordinate descent algorithm for complex-valued lasso and Gaussian graphical lasso, along with the corresponding \texttt{cxreg} R package, leverages the standard pathwise coordinate descent method by demonstrating its applicability to complex-valued settings [@deb2024regularized]. Our algorithm is defined on the realifying complex numbers via a ring isomorphism [e.g., @herstein1991topics; @dummit2004abstract]. Specifically, for a complex number $z\in\mathbb{C}$, we define the map:
-
-$$
-\varphi(z) = \begin{pmatrix}
-\mathbf{Re}(z) & -\mathbf{Im}(z) \\
-\mathbf{Im}(z) & \mathbf{Re}(z)
-\end{pmatrix}.
-$$
-We show that the $\varphi$ is a field isomorphism between $\mathbb{C}$ and the set
-
-$$
-\mathcal{M}^{2 \times 2} :=\left\{\begin{pmatrix} 
-a & - b \\ b & a
-\end{pmatrix} : a,b\in\mathbb{R}\right\}.
-$$
-
-Under this mapping, there exists connections between algebraic operations in the real and complex domains. We extend these operations to the $p$-dimensional case for $p>1$. Further details are provided in Section 3.1 of @deb2024regularized.
-
-We use this mapping to transform linear regression problems in the complex-valued setting into equivalent problems in the real-valued setting. The same mapping can then be applied to the Gaussian log-likelihood for complex-valued variables. For the $\ell_1$  penalty term used in lasso and graphical lasso, we exploit the fact that the modulus of each complex-valued coordinate is equivalent to the entrywise $\ell_2$ norm of its real and imaginary parts. This implies that the lasso formulation for $p$-dimensional complex-valued variables is equivalent to a $2p$-dimensional group lasso [@yuan2006model], where each group consists of the real and imaginary components of a single complex variable ($p$ groups in total).
-
-
-Finally, we implement the computational shortcuts, including *warm start* with the solutions for the sequence of the tuning parameter $\lambda$ in lasso and graphical lasso, and *active set selection* to reduce the number of predictors that need to be updated at each $\lambda$ [e.g., Chapter 5.4 in @hastie2015statistical]. In the R package, the user interface, such as the names of inputs for functions, is borrowed from the well-known pathwise coordinate descent R package \texttt{glmnet} [@friedman2010regularization]. A detailed vignette is available [Here](https://github.com/yk748/cxreg/blob/main/vignettes/cxreg.pdf).
-
-
-# Statement of need
-
[…]
