+++ R/buildPremat.default.R
@@ -1,13 +1,15 @@
 #' Build Prediction Matrix (Default Method)
 #'
-#' These are not intended for use by users. Constructs a prediction matrix for cross-validation folds, using the coefficient paths in \code{outlist}.
-#' Inspired by internal functions in the \code{glmnet} package.
+#' These are not intended for use by users. Constructs a prediction matrix for cross-validation folds, using the coefficient paths in \code{outlist}. Inspired by internal functions in the \pkg{glmnet} package.
 #' @param outlist A list of fitted models.
 #' @param lambda A vector of penalty values.
 #' @param x A predictor matrix.
 #' @param foldid Fold identifiers.
 #' @param alignment Alignment method (not used).
 #' @param \dots Other arguments
+#' @return A numeric matrix of predicted values with dimensions
+#'   \code{nrow(x)} by \code{length(lambda)}. Rows correspond to
+#'   observations and columns correspond to penalty values.
 #' @export
 buildPredmat.default <- function(outlist, lambda, x, foldid, alignment, ...){
 
@@ -46,6 +48,8 @@ buildPredmat.default <- function(outlist, lambda, x, foldid, alignment, ...){
 #' @param foldid Fold identifiers.
 #' @param alignment Alignment method (not used).
 #' @param \dots Other arguments
+#' @return The return value depends on the method; for the default method,
+#'   see \code{\link{buildPredmat.default}}.
 #' @export
 buildPredmat <- function(outlist, lambda, x, foldid, alignment, ...){
   UseMethod("buildPredmat")
