+++ paper.md
@@ -1,0 +1,150 @@
+# ---
+
+# title: 'cxreg: An R package for complex-valued lasso and graphical lasso'
+
+# tags:
+
+# &nbsp; - Spectral analysis,
+
+# &nbsp; - Regularization path,
+
+# &nbsp; - Coordinate descent ,
+
+# &nbsp; - lasso, 
+
+# &nbsp; - Graphical model
+
+# authors:
+
+# &nbsp; - name: Younghoon Kim
+
+# &nbsp;   orcid: 0009-0007-0117-5530
+
+# &nbsp;   equal-contrib: true
+
+# &nbsp;   affiliation: 1
+
+# &nbsp; - name: Navonil Deb
+
+# &nbsp;   equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
+
+# &nbsp;   affiliation: 1
+
+# &nbsp; - name: Sumanta Basu
+
+# &nbsp;   corresponding: true # (This is how to denote the corresponding author)
+
+# &nbsp;   affiliation: 1
+
+# affiliations:
+
+# &nbsp;- name: Department of Statistics and Data Science, Cornell University, United States
+
+# &nbsp;  index: 1
+
+# date: 19 July 2025
+
+# bibliography: cxreg-ref.bib
+
+# ---
+
+# 
+
+# \# Summary
+
+# 
+
+# \*Pathwise coordinate descent\* \[@friedman2007pathwise] is a popular iterative optimization method for solving penalized linear regression \[\*lasso\*; @tibshirani1996regression] and penalized Gaussian log-likelihood \[\*graphical lasso\*; e.g., @banerjee2008model] problems. Given a specified or derived sequence of tuning parameters $\\lambda$ in the penalty term, the algorithm updates one coordinate at a time by isolating the target variable using partial residuals.
+
+# 
+
+# The coordinate descent algorithm for complex-valued lasso and Gaussian graphical lasso, along with the corresponding \\texttt{cxreg} R package, leverages the standard pathwise coordinate descent method by demonstrating its applicability to complex-valued settings \[@deb2024regularized]. Our algorithm is defined on the realifying complex numbers via a ring isomorphism \[e.g., @herstein1991topics; @dummit2004abstract]. Specifically, for a complex number $z\\in\\mathbb{C}$, we define the map:
+
+# 
+
+# $$
+
+# \\varphi(z) = \\begin{pmatrix}
+
+# \\mathbf{Re}(z) \& -\\mathbf{Im}(z) \\\\
+
+# \\mathbf{Im}(z) \& \\mathbf{Re}(z)
+
+# \\end{pmatrix}.
+
+# $$
+
+# We show that the $\\varphi$ is a field isomorphism between $\\mathbb{C}$ and the set
+
+# 
+
+# $$
+
+# \\mathcal{M}^{2 \\times 2} :=\\left\\{\\begin{pmatrix} 
+
+# a \& - b \\\\ b \& a
+
+# \\end{pmatrix} : a,b\\in\\mathbb{R}\\right\\}.
+
+# $$
+
+# 
+
+# Under this mapping, there exists connections between algebraic operations in the real and complex domains. We extend these operations to the $p$-dimensional case for $p>1$. Further details are provided in Section 3.1 of @deb2024regularized.
+
+# 
+
[…]
