+++ vignette/cxreg.Rmd
@@ -1,0 +1,260 @@
+---
+title: "An Introduction to `cxreg`"
+author:
+  - Younghoon Kim
+  - Navonil Deb
+  - Sumanta Basu
+date: "`r format(Sys.time(), '%B %d, %Y')`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{An Introduction to cxreg}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+link-citations: true
+bibliography: cxreg_refs.bib
+---
+
+```{r include=FALSE}
+# the code in this chunk enables us to truncate the print output for each
+# chunk using the `out.lines` option
+# save the built-in output hook
+hook_output <- knitr::knit_hooks$get("output")
+
+# set a new output hook to truncate text output
+knitr::knit_hooks$set(output = function(x, options) {
+  if (!is.null(n <- options$out.lines)) {
+    x <- xfun::split_lines(x)
+    if (length(x) > n) {
+        
+      # truncate the output
+      x <- c(head(x, n), "....\n")
+    }
+    x <- paste(x, collapse = "\n")
+  }
+  hook_output(x, options)
+})
+```
+
+## Introduction
+
+\texttt{cxreg} is a package that fits complex-valued penalized regressions (Lasso) and Gaussian likelihoods (graphical Lasso) using an exact pathwise coordinate descent method. Similar to the well-known package glmnet [-@friedman2010regularization] for Lasso, cxreg computes the regularization path for complex-valued linear regression with a lasso penalty, referred to as \texttt{classo}, over a grid of values for the regularization parameter lambda. Likewise, to fit complex-valued Graphical Lasso models [-@friedman2008sparse], the regularization path is computed via \texttt{cglasso}. The package includes methods for prediction, cross-validation functions, printing and plotting utilities, as well as fitting for both model types.
+
+The authors of cxreg are Navonil Deb and Sumanta Basu, with contributions from Younghoon Kim. The R package is maintained by Younghoon Kim.
+
+This vignette describes basic usage of functions related to \texttt{classo} and \texttt{cglasso} models in 'cxreg' package in R. There is an original description of the algorithm that should be useful:
+
+* ["Regularized estimation of sparse spectral precision matrices"](https://arxiv.org/abs/2401.11128)
+
+
+\texttt{classo} solves the following problem: For $Y\in\mathbb{C}^n$ and $X\in\mathbb{C}^{n \times p}$,
+
+$$
+\min_{\beta \in \mathbb{C}^p} \frac{1}{2n} \|Y - X \beta\|_2^2 + \lambda\|\beta\|_1,
+$$
+
+over a grid of values of $\lambda$ covering the entire range of possible solutions. Here, since the absolute of complex-values means a complex modulus, $\beta\in\mathbb{C}^p$ so that the $\ell_1$-penalty can be viewed as
+
+$$
+\|\beta\|_1 = \sum_{j=1}^p |\beta_j| = \sum_{j=1}^p\|\mathrm{Re}(\beta_j) + \mathrm{Im}(\beta_j)\|_2,
+$$
+
+which is a group Lasso with $p$ groups, each of size 2.
+
+
[…]
