+++ paper.bib
@@ -1,29 +1,76 @@
-@article{friedman2021package,
-  title={Package ‘glmnet’},
-  author={Friedman, Jerome and Hastie, Trevor and Tibshirani, Rob and Narasimhan, Balasubramanian and Tay, Kenneth and Simon, Noah and Qian, Junyang},
-  journal={CRAN R Repositary},
-  volume={595},
-  year={2021}
+@article{friedman2007pathwise,
+  title={Pathwise coordinate optimization},
+  author={Friedman, Jerome and Hastie, Trevor and H{\"o}fling, Holger and Tibshirani, Robert},
+  journal={The Annals of Applied Statistics},
+  pages={302--332},
+  year={2007},
+  publisher={JSTOR}
 }
-
-@article{friedman2010regularization,
-  title={Regularization paths for generalized linear models via coordinate descent},
-  author={Friedman, Jerome and Hastie, Trevor and Tibshirani, Rob},
-  journal={Journal of Statistical Software},
-  volume={33},
+@article{tibshirani1996regression,
+  title={Regression shrinkage and selection via the lasso},
+  author={Tibshirani, Robert},
+  journal={Journal of the Royal Statistical Society Series B: Statistical Methodology},
+  volume={58},
   number={1},
-  pages={1},
-  year={2010},
-  publisher={NIH Public Access}
+  pages={267--288},
+  year={1996},
+  publisher={Oxford University Press}
 }
-
+@article{banerjee2008model,
+  title={Model selection through sparse maximum likelihood estimation for multivariate Gaussian or binary data},
+  author={Banerjee, Onureena and El Ghaoui, Laurent and d'Aspremont, Alexandre},
+  journal={The Journal of Machine Learning Research},
+  volume={9},
+  pages={485--516},
+  year={2008},
+  publisher={JMLR. org}
+}
 @article{deb2024regularized,
   title={Regularized estimation of sparse spectral precision matrices},
   author={Deb, Navonil and Kuceyeski, Amy and Basu, Sumanta},
   journal={arXiv preprint arXiv:2401.11128},
   year={2024}
 }
-
+@book{herstein1991topics,
+  title={Topics in Agebra},
+  author={Herstein, Israel N},
+  year={1991},
+  publisher={John Wiley \& Sons}
+}
+@book{dummit2004abstract,
+  title={Abstract Algebra},
+  author={Dummit, David Steven and Foote, Richard M and others},
+  volume={3},
+  year={2004},
+  publisher={Wiley Hoboken}
+}
+@article{yuan2006model,
+  title={Model selection and estimation in regression with grouped variables},
+  author={Yuan, Ming and Lin, Yi},
+  journal={Journal of the Royal Statistical Society Series B: Statistical Methodology},
+  volume={68},
+  number={1},
+  pages={49--67},
+  year={2006},
+  publisher={Oxford University Press}
+}
+@article{friedman2010regularization,
+  title={Regularization paths for generalized linear models via coordinate descent},
+  author={Friedman, Jerome and Hastie, Trevor and Tibshirani, Rob},
+  journal={Journal of statistical software},
+  volume={33},
+  pages={1--22},
+  year={2010}
+}
+@article{hastie2015statistical,
+  title={Statistical Learning with Sparsity},
+  author={Hastie, Trevor and Tibshirani, Robert and Wainwright, Martin},
+  journal={Monographs on Statistics and Applied Probability},
+  volume={143},
+  number={143},
+  pages={8},
+  year={2015}
+}
 @article{friedman2008sparse,
   title={Sparse inverse covariance estimation with the graphical lasso},
   author={Friedman, Jerome and Hastie, Trevor and Tibshirani, Robert},
[…]
