+++ paper.md
@@ -1,150 +1,181 @@
-# ---
+# Summary
+
+*Pathwise coordinate descent* (Friedman et al. 2007) is a popular
+iterative optimization method for solving penalized linear regression
+\[*lasso*; Tibshirani (1996)\] and penalized Gaussian log-likelihood
+\[*graphical lasso*; e.g., Banerjee, El Ghaoui, and d’Aspremont (2008)\]
+problems. Given a specified or derived sequence of tuning parameters *λ*
+in the penalty term, the algorithm updates one coordinate at a time by
+isolating the target variable using partial residuals.
+
+The coordinate descent algorithm for complex-valued lasso and Gaussian
+graphical lasso, along with the corresponding R package, leverages the
+standard pathwise coordinate descent method by demonstrating its
+applicability to complex-valued settings (Deb, Kuceyeski, and Basu
+2024). Our algorithm is defined on the realifying complex numbers via a
+ring isomorphism (e.g., Herstein 1991; Dummit, Foote, et al. 2004).
+Specifically, for a complex number *z* ∈ ℂ, we define the map:
+
+$$
+\varphi(z) = \begin{pmatrix}
+\mathbf{Re}(z) & -\mathbf{Im}(z) \\
+\mathbf{Im}(z) & \mathbf{Re}(z)
+\end{pmatrix}.
+$$
+We show that the *φ* is a field isomorphism between ℂ and the set
+
+$$
+\mathcal{M}^{2 \times 2} :=\left\\\begin{pmatrix} 
+a & - b \\ b & a
+\end{pmatrix} : a,b\in\mathbb{R}\right\\.
+$$
+
+Under this mapping, there exists connections between algebraic
+operations in the real and complex domains. We extend these operations
+to the *p*-dimensional case for *p* &gt; 1. Further details are provided
+in Section 3.1 of Deb, Kuceyeski, and Basu (2024).
+
+We use this mapping to transform linear regression problems in the
+complex-valued setting into equivalent problems in the real-valued
+setting. The same mapping can then be applied to the Gaussian
+log-likelihood for complex-valued variables. For the ℓ<sub>1</sub>
+penalty term used in lasso and graphical lasso, we exploit the fact that
+the modulus of each complex-valued coordinate is equivalent to the
+entrywise ℓ<sub>2</sub> norm of its real and imaginary parts. This
+implies that the lasso formulation for *p*-dimensional complex-valued
+variables is equivalent to a 2*p*-dimensional group lasso (Yuan and Lin
+2006), where each group consists of the real and imaginary components of
+a single complex variable (*p* groups in total).
+
+Finally, we implement the computational shortcuts, including *warm
+start* with the solutions for the sequence of the tuning parameter *λ*
+in lasso and graphical lasso, and *active set selection* to reduce the
+number of predictors that need to be updated at each *λ* (e.g., Chapter
+5.4 in Hastie, Tibshirani, and Wainwright 2015). In the R package, the
+user interface, such as the names of inputs for functions, is borrowed
+from the well-known pathwise coordinate descent R package (Friedman,
+Hastie, and Tibshirani 2010). A detailed vignette is available
+[Here](https://github.com/yk748/cxreg/blob/main/vignettes/cxreg.pdf).
+
+# Statement of need
+
+Solving complex-valued penalized Gaussian likelihood (as well as
+penalized linear regression as a part of the problem) is critical in
+high-dimensional time series analysis, as examining partial spectral
+coherence in the frequency domain is analogous to examining partial
+correlations in Gaussian graphical models (e.g., Priestley 1988;
+Dahlhaus 2000). With the development of the local Whittle likelihood
+approximation (Whittle 1951), the computation of the inverse spectral
[…]
