+++ paper.md
@@ -4,9 +4,9 @@
 iterative optimization method for solving penalized linear regression
 \[*lasso*; Tibshirani (1996)\] and penalized Gaussian log-likelihood
 \[*graphical lasso*; e.g., Banerjee, El Ghaoui, and d’Aspremont (2008)\]
-problems. Given a specified or derived sequence of tuning parameters *λ*
-in the penalty term, the algorithm updates one coordinate at a time by
-isolating the target variable using partial residuals.
+problems. Given a specified or derived sequence of tuning parameters
+$\lambda$ in the penalty term, the algorithm updates one coordinate at a
+time by isolating the target variable using partial residuals.
 
 The coordinate descent algorithm for complex-valued lasso and Gaussian
 graphical lasso, along with the corresponding R package, leverages the
@@ -14,47 +14,48 @@ standard pathwise coordinate descent method by demonstrating its
 applicability to complex-valued settings (Deb, Kuceyeski, and Basu
 2024). Our algorithm is defined on the realifying complex numbers via a
 ring isomorphism (e.g., Herstein 1991; Dummit, Foote, et al. 2004).
-Specifically, for a complex number *z* ∈ ℂ, we define the map:
+Specifically, for a complex number $z\in\mathbb{C}$, we define the map:
 
 $$
 \varphi(z) = \begin{pmatrix}
 \mathbf{Re}(z) & -\mathbf{Im}(z) \\
 \mathbf{Im}(z) & \mathbf{Re}(z)
 \end{pmatrix}.
-$$
-We show that the *φ* is a field isomorphism between ℂ and the set
+$$ We show that the $\varphi$ is a field isomorphism between
+$\mathbb{C}$ and the set
 
 $$
-\mathcal{M}^{2 \times 2} :=\left\\\begin{pmatrix} 
+\mathcal{M}^{2 \times 2} :=\left\{\begin{pmatrix} 
 a & - b \\ b & a
-\end{pmatrix} : a,b\in\mathbb{R}\right\\.
+\end{pmatrix} : a,b\in\mathbb{R}\right\}.
 $$
 
 Under this mapping, there exists connections between algebraic
 operations in the real and complex domains. We extend these operations
-to the *p*-dimensional case for *p* &gt; 1. Further details are provided
-in Section 3.1 of Deb, Kuceyeski, and Basu (2024).
+to the $p$-dimensional case for $p>1$. Further details are provided in
+Section 3.1 of Deb, Kuceyeski, and Basu (2024).
 
 We use this mapping to transform linear regression problems in the
 complex-valued setting into equivalent problems in the real-valued
 setting. The same mapping can then be applied to the Gaussian
-log-likelihood for complex-valued variables. For the ℓ<sub>1</sub>
-penalty term used in lasso and graphical lasso, we exploit the fact that
-the modulus of each complex-valued coordinate is equivalent to the
-entrywise ℓ<sub>2</sub> norm of its real and imaginary parts. This
-implies that the lasso formulation for *p*-dimensional complex-valued
-variables is equivalent to a 2*p*-dimensional group lasso (Yuan and Lin
-2006), where each group consists of the real and imaginary components of
-a single complex variable (*p* groups in total).
+log-likelihood for complex-valued variables. For the $\ell_1$ penalty
+term used in lasso and graphical lasso, we exploit the fact that the
+modulus of each complex-valued coordinate is equivalent to the entrywise
[…]
