### 2025-08-09 #8668 @yk748
Dear @crvernon, we converted the Rmd file to an md file, and the document’s title changed to paper.md. Please review our resubmission. Thank you for your time and consideration.
