+++ vignette/cxreg.Rmd
@@ -5,7 +5,7 @@ author:
   - Navonil Deb
   - Sumanta Basu
 date: "`r format(Sys.time(), '%B %d, %Y')`"
-output: rmarkdown::html_vignette
+output: pdf_document
 vignette: >
   %\VignetteIndexEntry{An Introduction to cxreg}
   %\VignetteEngine{knitr::rmarkdown}
