+++ paper.md
@@ -49,7 +49,7 @@ Under this mapping, there exists connections between algebraic operations in the
 We use this mapping to transform linear regression problems in the complex-valued setting into equivalent problems in the real-valued setting. The same mapping can then be applied to the Gaussian log-likelihood for complex-valued variables. For the $\ell_1$  penalty term used in lasso and graphical lasso, we exploit the fact that the modulus of each complex-valued coordinate is equivalent to the entrywise $\ell_2$ norm of its real and imaginary parts. This implies that the lasso formulation for $p$-dimensional complex-valued variables is equivalent to a $2p$-dimensional group lasso [@yuan2006model], where each group consists of the real and imaginary components of a single complex variable ($p$ groups in total).
 
 
-Finally, we implement the computational shortcuts, including *warm start* with the solutions for the sequence of the tuning parameter $\lambda$ in lasso and graphical lasso, and *active set selection* to reduce the number of predictors that need to be updated at each $\lambda$ [e.g., Chapter 5.4 in @hastie2015statistical]. In the R package, the user interface, such as the names of inputs for functions, is borrowed from the well-known pathwise coordinate descent R package \texttt{glmnet} [@friedman2010regularization]. A detailed vignette is available [Here](https://github.com/yk748/cxreg/blob/main/vignettes/cxreg.pdf).
+Finally, we implement the computational shortcuts, including *warm start* with the solutions for the sequence of the tuning parameter $\lambda$ in lasso and graphical lasso, and *active set selection* to reduce the number of predictors that need to be updated at each $\lambda$ [e.g., Chapter 5.4 in @hastie2015statistical]. In the R package, the user interface, such as the names of inputs for functions, is borrowed from the well-known pathwise coordinate descent R package \texttt{glmnet} [@friedman2010regularization]. A detailed vignette is available [Here](https://github.com/yk748/cxreg/blob/main/vignette/cxreg.pdf).
 
 
 # Statement of need
