+++ paper.bib
@@ -37,13 +37,7 @@
   year={1991},
   publisher={John Wiley \& Sons}
 }
-@book{dummit2004abstract,
-  title={Abstract Algebra},
-  author={Dummit, David Steven and Foote, Richard M and others},
-  volume={3},
-  year={2004},
-  publisher={Wiley Hoboken}
-}
+
 @article{yuan2006model,
   title={Model selection and estimation in regression with grouped variables},
   author={Yuan, Ming and Lin, Yi},
@@ -87,16 +81,7 @@
   year={1988},
   publisher={Oxford University Press}
 }
-@article{dahlhaus2000graphical,
-  title={Graphical interaction models for multivariate time series1},
-  author={Dahlhaus, Rainer},
-  journal={Metrika},
-  volume={51},
-  number={2},
-  pages={157--172},
-  year={2000},
-  publisher={Springer}
-}
+
 @article{whittle1951hypothesis,
   title={Hypothesis Testing in Time Series Analysis},
   author={Whittle, Peter},
