### 2025-08-22 #8668 @crvernon
@yk748 - I am not sure I understand your statement:  "We have fixed the hyperlink to the PDF vignette. All functions that need to be tested are included there."

Are there tests that can be ran locally via `testthat` or a similar library in your case that cover the functionality of your code so that when a user modifies your code it still produces the desired outcome?  Preferably these would be automated (e.g., tests that run via GitHub Actions, etc. upon push to the repository).
