### 2025-08-22 #8668 @yk748
@crvernon Thank you for your follow-up. The package is based on a subset of the authors’ work (arxiv.org/pdf/2401.11128), which is also cited as a reference in this JOSS paper. The inclusive simulations comparing our method to benchmarks are already included there.
We also followed the format of the subset of the authors’ previous JOSS publication (joss.theoj.org/papers/10.21105/joss.01077.pdf
). In line with that format, all implementations are provided in the hyperlinked vignette.
