### 2025-09-05 #8668 @crvernon
> [@yk748](https://github.com/yk748)
>
> > The inclusive simulations comparing our method to benchmarks are already included there.
>
> What I am asking for are tests (e.g., unit, integration, etc.) that accompany your code base and not just the validation and benchmarking.
>
> Note in the 2018 paper you referenced above (joss.theoj.org/papers/10.21105/joss.01077.pdf) the supporting repository had a test suite as well (see https://github.com/sumbose/iRF/tree/master/tests).

@yk748 - just posting a reminder of the above inquiry
