+++ paper.md
@@ -33,9 +33,21 @@ bibliography: paper.bib
 
 # Summary
 
-*Pathwise coordinate descent* [@friedman2007pathwise] is a popular iterative optimization method for solving penalized linear regression [*lasso*; @tibshirani1996regression] and penalized Gaussian log-likelihood [*graphical lasso*; e.g., @banerjee2008model] problems. Given a specified or derived sequence of tuning parameters $\lambda$ in the penalty term, the algorithm updates one coordinate at a time by isolating the target variable using partial residuals.
+*Pathwise coordinate descent* (Friedman et al. 2007) is a popular
+iterative optimization method for solving penalized linear regression
+\[*lasso*; Tibshirani (1996)\] and penalized Gaussian log-likelihood
+\[*graphical lasso*; e.g., Banerjee, El Ghaoui, and d’Aspremont (2008)\]
+problems. Given a specified or derived sequence of tuning parameters
+$\lambda$ in the penalty term, the algorithm updates one coordinate at a
+time by isolating the target variable using partial residuals.
 
-The coordinate descent algorithm for complex-valued lasso and Gaussian graphical lasso, along with the corresponding `cxreg' R package, leverages the standard pathwise coordinate descent method by demonstrating its applicability to complex-valued settings [@deb2024regularized]. Our algorithm is defined on the realifying complex numbers via a ring isomorphism [e.g., @herstein1991topics]. Specifically, for a complex number $z\in\mathbb{C}$, we define the map:
+The coordinate descent algorithm for complex-valued lasso and Gaussian
+graphical lasso, along with the corresponding \`cxreg’ R package,
+leverages the standard pathwise coordinate descent method by
+demonstrating its applicability to complex-valued settings (Deb,
+Kuceyeski, and Basu 2024). Our algorithm is defined on the realifying
+complex numbers via a ring isomorphism (e.g., Herstein 1991).
+Specifically, for a complex number $z\in\mathbb{C}$, we define the map:
 
 $$
 \varphi(z) =
@@ -43,8 +55,8 @@ $$
 \mathbf{Re}(z) & -\mathbf{Im}(z) \\
 \mathbf{Im}(z) &  \mathbf{Re}(z)
 \end{pmatrix}.
-$$
-We show that the $\varphi$ is a field isomorphism between $\mathbb{C}$ and the set
+$$ We show that the $\varphi$ is a field isomorphism between
+$\mathbb{C}$ and the set
 
 $$
 \mathcal{M}^{2 \times 2} :=\biggl\{\begin{pmatrix} 
@@ -52,31 +64,215 @@ a & - b \\ b & a
 \end{pmatrix} : a,b\in\mathbb{R}\biggr\}.
 $$
 
-Under this mapping, there exists connections between algebraic operations in the real and complex domains. We extend these operations to the $p$-dimensional case for $p>1$. Further details are provided in Section 3.1 of @deb2024regularized.
+Under this mapping, there exists connections between algebraic
+operations in the real and complex domains. We extend these operations
+to the $p$-dimensional case for $p>1$. Further details are provided in
+Section 3.1 of Deb, Kuceyeski, and Basu (2024).
 
[…]
