+++ R/plot.classo.R
@@ -22,8 +22,8 @@
 #'
 #' @method plot classo
 #' @export
-plot.classo <- function(x, xvar=c("norm","lambda"),label=FALSE,...){
+plot.classo <- function(x, xvar=c("norm","lambda","dev"),label=FALSE,...){
 
-  # xvar <- match.arg(xvar)
+  xvar <- match.arg(xvar)
   plotCoef(x$beta,lambda=x$lambda,df=x$df,dev=x$dev.ratio,label=label,xvar=xvar,...)
 }
