+++ paper.md
@@ -33,112 +33,50 @@ bibliography: paper.bib
 
 # Summary
 
-*Pathwise coordinate descent* (Friedman et al. 2007) is a popular
-iterative optimization method for solving penalized linear regression
-\[*lasso*; Tibshirani (1996)\] and penalized Gaussian log-likelihood
-\[*graphical lasso*; e.g., Banerjee, El Ghaoui, and d’Aspremont (2008)\]
-problems. Given a specified or derived sequence of tuning parameters
-$\lambda$ in the penalty term, the algorithm updates one coordinate at a
-time by isolating the target variable using partial residuals.
+*Pathwise coordinate descent* (Friedman et al. 2007) is a widely used iterative optimization method for solving penalized linear regression \[*lasso*; Tibshirani (1996)\] and penalized Gaussian log-likelihood \[*graphical lasso*; e.g., Banerjee, El Ghaoui, and d’Aspremont (2008)\] problems. For a given sequence of tuning parameters $\lambda$ in the penalty term, the algorithm updates one coefficient at a time by computing partial residuals.
 
-The coordinate descent algorithm for complex-valued lasso and Gaussian
-graphical lasso, along with the corresponding \`cxreg’ R package,
-leverages the standard pathwise coordinate descent method by
-demonstrating its applicability to complex-valued settings (Deb,
-Kuceyeski, and Basu 2024). Our algorithm is defined on the realifying
-complex numbers via a ring isomorphism (e.g., Herstein 1991).
-Specifically, for a complex number $z\in\mathbb{C}$, we define the map:
+The coordinate descent algorithm for complex-valued lasso and Gaussian graphical lasso, implemented in the cxreg R package, extends the standard pathwise coordinate descent method to complex-valued data (Deb, Kuceyeski, and Basu 2024). The algorithm operates by *realifying* complex numbers through a ring isomorphism (Herstein 1991). Specifically, for a complex number $z \in \mathbb{C}$, we define the map:
 
 $$
-\varphi(z) =
-\begin{pmatrix}
-\mathrm{Re}(z) & -\mathrm{Im}(z) \\
-\mathrm{Im}(z) &  \mathrm{Re}(z)
+\varphi(z) = 
+\begin{pmatrix} 
+\mathrm{Re}(z) & -\mathrm{Im}(z) \\ 
+\mathrm{Im}(z) & \mathrm{Re}(z) 
 \end{pmatrix}.
-$$ We show that the $\varphi$ is a field isomorphism between
+$$ 
+
+We show that the $\varphi$ is a field isomorphism between
 $\mathbb{C}$ and the set
 
 $$
-\mathcal{M}^{2 \times 2} :=\left\{\begin{pmatrix} 
-a & - b \\ b & a
-\end{pmatrix} : a,b\in\mathbb{R}\right\}.
+\mathcal{M}^{2 \times 2} := \left\{ 
+\begin{pmatrix}
+a & -b \\ 
+b & a
+\end{pmatrix} 
+: a,b \in \mathbb{R} 
+\right\}.
 $$
 
-Under this mapping, there exists connections between algebraic
-operations in the real and complex domains. We extend these operations
-to the $p$-dimensional case for $p>1$. Further details are provided in
-Section 3.1 of Deb, Kuceyeski, and Basu (2024).
+Under this mapping, algebraic operations in the complex domain correspond directly to those in the real domain. We extend these operations to the $p$-dimensional case $(p>1)$; see Section 3.1 of Deb, Kuceyeski, and Basu (2024) for details.
 
-We use this mapping to transform linear regression problems in the
-complex-valued setting into equivalent problems in the real-valued
-setting. The same mapping can then be applied to the Gaussian
-log-likelihood for complex-valued variables. For the $\ell_{1}$ penalty
[…]
