+++ src/cglassocd_scaled.f
@@ -169,4 +169,4 @@ c     cglasso with covariate scaling/covariance adapted penalty
       return
       end
             
-                     
+                  
