+++ DESCRIPTION
@@ -29,6 +29,8 @@ Suggests:
     knitr,
     rmarkdown,
     testthat (>= 3.0.0),
+    tinytex,
+    tools,
     xfun
 VignetteBuilder: 
     knitr
