+++ paper.md
@@ -50,13 +50,13 @@ $\mathbb{C}$ and the set
 
 $$
 \mathcal{M}^{2 \times 2} := 
-\left\{ 
+\{ 
 \begin{pmatrix}
 a & -b \\ 
 b & a
 \end{pmatrix} 
 : a,b \in \mathbb{R} 
-\right\}.
+\}.
 $$
 
 Under this mapping, algebraic operations in the complex domain correspond directly to those in the real domain. We extend these operations to the $p$-dimensional case $(p>1)$; see Section 3.1 of Deb, Kuceyeski, and Basu (2024) for details.
