### 2025-12-28 #9698 @editorialbot
[REVIEW]: cxreg: An R package for complex-valued lasso and graphical lasso
**Submitting author:** @yk748 (<a href="http://orcid.org/0009-0007-0117-5530">Younghoon Kim</a>)
**Repository:** https://github.com/yk748/cxreg
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
