### 2026-01-28 #9698 @tambstat
@gkthiruvathukal Some comments below.

1. The R package is well-written with all necessary instructions included to run the functions.
2. The functions run without any errors and also runs in very reasonable time, even for higher dimensional inputs.
3. When installing the package on Windows via the _devtools_ route, a correct version of _Rtools_ has to be installed first. Some guidance on this can be made available in the installation description.
4. The main application considered in the work Deb et. al (2024) is sparse spectral or inverse spectral matrix estimation using the LASSO penalization approach. Including a simulation example or function to illustrate this would be quite helpful. I believe this would make the package even more attractive for folks in disciplines beyond statistics.
