### 2026-02-25 #9698 @gkthiruvathukal
Clarification: An incomplete checklist is a blocker for acceptance.
