### 2026-03-04 #9698 @gkthiruvathukal
@tambstat also has unchecked items.
