### 2026-04-13 #9698 @gkthiruvathukal
@yk748 The main blocker is https://github.com/openjournals/joss-reviews/issues/9698#issuecomment-3999538812.

I also need @adallak and @tambstat to confirm that all issues have been resolved. But both review checklists must be complete with no unchecked items.
