### 2026-04-13 #9698 @yk748
@gkthiruvathukal  Thank you for the follow-up. I now understand the system to some extent. If I understand correctly, all the items in the checklist must be completed. The unchecked items are not due to reviewers’ mistakes; rather, they indicate that the corresponding code or paper for those aspects needs to be updated by the authors. Also, these updates can be made by the authors at any time without limitation. Is this correct?
