### 2026-04-16 #9698 @adallak
> [@yk748](https://github.com/yk748) The main blocker is [#9698 (comment)](https://github.com/openjournals/joss-reviews/issues/9698#issuecomment-3999538812).
>
> I also need [@adallak](https://github.com/adallak) and [@tambstat](https://github.com/tambstat) to confirm that all issues have been resolved. But both review checklists must be complete with no unchecked items.

@gkthiruvathukal Is there an updated version of the paper? It seems https://github.com/yk748/cxreg contains the old version.
