+++ paper.bib
@@ -32,7 +32,7 @@
   year={2024}
 }
 @book{herstein1991topics,
-  title={Topics in Agebra},
+  title={Topics in Algebra},
   author={Herstein, Israel N},
   year={1991},
   publisher={John Wiley \& Sons}
@@ -132,3 +132,42 @@
   year={2023},
   publisher={Institute of Mathematical Statistics}
 }
+
+
+@article{dallakyan2024learning,
+  title={On learning time series {DAG}s: A frequency domain approach},
+  author={Dallakyan, Aramayis},
+  journal={Econometrics and Statistics},
+  year={2024},
+  publisher={Elsevier}
+}
+
+
+@article{dallakyan2022time,
+  title={Time series graphical lasso and sparse {VAR} estimation},
+  author={Dallakyan, Aramayis and Kim, Rakheon and Pourahmadi, Mohsen},
+  journal={Computational Statistics \& Data Analysis},
+  volume={176},
+  pages={107557},
+  year={2022},
+  publisher={Elsevier}
+}
+
+@book{artin2018algebra,
+  title={Algebra},
+  author={Artin, Michael},
+  series={Pearson Modern Classics for Advanced Mathematics Series},
+  year={2018},
+  publisher={Pearson}
+}
+
+
+
+
+
+
+
+
+
+
+
