+++ paper.md
@@ -4,7 +4,7 @@ tags:
   - Spectral analysis
   - Regularization path
   - Coordinate descent 
-  - lasso
+  - Lasso
   - Graphical model
 authors:
   - name: Younghoon Kim
@@ -13,29 +13,25 @@ authors:
     affiliation: 1
   - name: Navonil Deb
     equal-contrib: true 
-    affiliation: 1
+    affiliation: 2
   - name: Sumanta Basu
     corresponding: true 
-    affiliation: 1
+    affiliation: 2
 affiliations:
- - name: Department of Statistics and Data Science, Cornell University, United States
+ - name: Center for Data Science for Enterprise and Society, Cornell University, United States
    index: 1
-date: 2025-08-14
-output:
-  md_document:
-    variant: gfm         
-    preserve_yaml: true  
-  pdf_document:          
-    number_sections: true
+ - name: Department of Statistics and Data Science, Cornell University, United States
+   index: 2
+date: 2026-04-20
 citeproc: true
 bibliography: paper.bib
 ---
 
 # Summary
 
-*Pathwise coordinate descent* (Friedman et al. 2007) is a widely used iterative optimization method for solving penalized linear regression \[*lasso*; Tibshirani (1996)\] and penalized Gaussian log-likelihood \[*graphical lasso*; e.g., Banerjee, El Ghaoui, and d’Aspremont (2008)\] problems. For a given sequence of tuning parameters $\lambda$ in the penalty term, the algorithm updates one coefficient at a time by computing partial residuals.
+*Pathwise coordinate descent* [@friedman2007pathwise] is a widely used iterative optimization method for solving penalized linear regression (*lasso*; [@tibshirani1996regression]) and penalized Gaussian log-likelihood (*graphical lasso*; e.g., Banerjee, El Ghaoui, and d’Aspremont [@banerjee2008model]) problems. For a given sequence of tuning parameters $\lambda$ in the penalty term, the algorithm updates one coefficient at a time by computing partial residuals.
 
-The coordinate descent algorithm for complex-valued lasso and Gaussian graphical lasso, implemented in the cxreg R package, extends the standard pathwise coordinate descent method to complex-valued data (Deb, Kuceyeski, and Basu 2024). The algorithm operates by *realifying* complex numbers through a ring isomorphism (Herstein 1991). Specifically, for a complex number $z \in \mathbb{C}$, we define the map:
+The coordinate descent algorithm for complex-valued lasso and Gaussian graphical lasso, implemented in the cxreg R package, extends the standard pathwise coordinate descent method to complex-valued data (Deb, Kuceyeski, and Basu [@deb2024regularized]). The algorithm operates by *realifying* complex numbers through a ring isomorphism [@herstein1991topics]. Specifically, for a complex number $z \in \mathbb{C}$, we define the map:
 
 $$
 \varphi(z) = 
@@ -45,8 +41,8 @@ $$
 \end{pmatrix}.
 $$ 
 
-We show that the $\varphi$ is a field isomorphism between
-$\mathbb{C}$ and the set $\mathcal{M}^{2 \times 2}$ whose entries consists of $a,b \in \mathbb{R}$ to form 2 by 2 matrices,
+It is known that $\varphi$ is a field isomorphism between
+$\mathbb{C}$ and the set $\mathcal{M}^{2 \times 2}$ whose entries consists of $a,b \in \mathbb{R}$ to form 2 by 2 matrices [@artin2018algebra],
 
 $$
 \begin{pmatrix}
@@ -55,159 +51,29 @@ b & a
 \end{pmatrix}.
 $$
 
[…]
