+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: cxreg
 Title: Complex-Valued Lasso and Complex-Valued Graphical Lasso
-Version: 1.0.0
+Version: 1.0.5
 Authors@R: c(
     person("Younghoon", "Kim", , "yk748@cornell.edu", role = c("aut", "cre")),
     person("Navonil", "Deb", role = "aut"),

[…]
