### 2026-04-21 #9698 @yk748
@adallak Thank you for your comments. I appreciate your interest and feedback. Please find my responses below. Your comments are in italics.

_Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support_

Re: For Community guidelines, I added the section “Reporting Issues” in the README to describe how to report bugs and request features via GitHub Issues, including guidance on providing reproducible examples and session information.

_State of the field: Do the authors describe how this software compares to other commonly-used packages?_

Re: Thank you for the suggestion. We have revised the State of the field section to include references to related work and existing software, including tsglasso (Dallakyan et al., 2021) and FreDom (Dallakyan, 2024). We also clarified how our approach differs and added references to benchmark comparisons in Deb, Kuceyeski, and Basu (2024).

_References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?_

Re: Thank you for reviewing this checklist item. The reference list is complete, and all works are appropriately cited.

_Text-related comments: In lines 17-21, the paper states "We show that the ... field isomorphism ..." I believe this is a classical, well-known fact (see, for example, Arin 2011). Similarly, if my memory is correct, the extension to the p-dimensional case (p>1) for a ring homomorphism is also a classical result with a componentwise multiplication operator._

Re: We agreed that the field isomorphism result and its $p$-dimensional extension are classical facts. We have revised the text to reflect this by replacing "We show that" with "It is well known that," and have added appropriate references to acknowledge the classical nature of these results.

_The paper is missing references to various works on complex-valued graphical models and existing software packages. For example, the R package tsglasso from Dallakyan et al. (2021) and the FreDom in Python (Dallakyan, 2024), where a similar closed-form solution is derived using Wirtinger calculus._

Re: Please refer to the reply to State of the field.

_Package-related: I didn't have any issues running the examples, but the package is missing expected sanity checks. For example, it doesn't require that the inputs be complex-valued, which results in an unexpected error. For classo, it is not required that y be a one-dimensional vector. I suggest that the authors include more automated tests and sanity checks to improve the package._

Re: Thank you for this helpful feedback. We agreed that the package would benefit from more robust input validation. Specifically, we added the following sanity checks:
1. A check that all inputs are complex-valued, returning an informative error message if not.
2. A check in classo that the y has a proper format.
3. Additional automated tests to cover these and other edge cases.
