### 2026-04-21 #9698 @yk748
@tambstat  I am grateful to you for your constructive comments. Please find my responses below. Your comments are in italics.

_Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item._

Re: Thank you for reviewing this checklist item. This paper contains no original data. The data application example uses a publicly accessible dataset, and instructions on how to access it are provided.

_Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item._

Re: Thank you for reviewing this checklist item. All results presented in the paper are fully reproducible. We provide a detailed vignette that contains instructions and code to reproduce all examples and results. The vignette is available at: https://github.com/yk748/cxreg/blob/main/doc/cxreg.pdf

_Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?_

Re: Thank you for reviewing this checklist item. The package currently includes basic automated tests. However, in light of the reviewer's earlier comment regarding missing sanity checks, we expanded the test suite to include more comprehensive input validation. We will also update the repository accordingly as we receive the reports from users.

_Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support_

Re: Please refer to the comment made to @adallak.

_State of the field: Do the authors describe how this software compares to other commonly-used packages?_

Re: Please refer to the comment made to @adallak.

_References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?_

Re: Please refer to the comment made to @adallak.

_The R package is well-written with all necessary instructions included to run the functions. The functions run without any errors and also runs in very reasonable time, even for higher dimensional inputs._

Re: Thank you for the positive feedback. We are glad to hear that the package runs without errors and performs well.

_When installing the package on Windows via the devtools route, a correct version of Rtools has to be installed first. Some guidance on this can be made available in the installation description._

Re: We agreed that Windows users may need to install Rtools before installing the package via devtools. We will add a note to the section `cxreg` installation for Windows in README.

_The main application considered in the work Deb et. al (2024) is sparse spectral or inverse spectral matrix estimation using the LASSO penalization approach. Including a simulation example or function to illustrate this would be quite helpful. I believe this would make the package even more attractive for folks in disciplines beyond statistics._

Re: Thank you for the suggestion. We agreed that including a simulation example illustrating sparse spectral precision matrix estimation would make the package more accessible to a broader audience. We added a dedicated example in the vignette that demonstrates the full workflow, from generating complex-valued discrete Fourier transformation from the time series data to estimating the sparse spectral precision matrix to help users from disciplines beyond statistics better understand the practical application of the package.
