+++ doc/cxreg.Rmd
@@ -14,12 +14,18 @@ link-citations: true
 bibliography: cxreg_refs.bib
 ---
 
-```{r setup-pdf, include=FALSE}
-knitr::opts_chunk$set(
-  fig.width  = 5,
-  fig.height = 4,
-  dpi        = 96      
-)
+```{r include=FALSE}
+hook_output <- knitr::knit_hooks$get("output")
+knitr::knit_hooks$set(output = function(x, options) {
+  if (!is.null(n <- options$out.lines)) {
+    x <- xfun::split_lines(x)
+    if (length(x) > n) {
+      x <- c(head(x, n), "....\n")
+    }
+    x <- paste(x, collapse = "\n")
+  }
+  hook_output(x, options)
+})
 ```
 
 ## Introduction
@@ -288,13 +294,65 @@ plot(fit_cglasso_II$Theta_list,index=fit_cglasso_II$min_index,type="imaginary",l
 This concludes the basic usage of \texttt{cglasso}.
 
 
-```{r compact-pdf, include=FALSE, eval=TRUE}
-tryCatch({
-  Sys.setenv(R_GSCMD = "C:/Program Files/gs/gs9.56.1/bin/gswin64c.exe")
-  f <- tools::file_path_as_absolute("cxreg.pdf")
-  if (file.exists(f)) {
-    tools::compactPDF(paths = f, gs_quality = "ebook")
-  }
-}, error = function(e) invisible(NULL))
+## Example: Sparse Spectral Precision Matrix Estimation
+
+As the real-data example from the time series, the workflow using this package should cover these steps:
+
+### Step 1 — Generate a multivariate Gaussian white noise process:
+
+```{r}
+library(mvtnorm)
+set.seed(1010)
+p <- 10    # number of variables
+n <- 500   # sample size
+
+# sparse precision matrix (tridiagonal)
+C <- diag(0.7, p)
+C[row(C) == col(C) + 1] <- 0.3
+C[row(C) == col(C) - 1] <- 0.3
+Sigma <- solve(C)
+
+# generate multivariate Gaussian time series
+X_t <- rmvnorm(n = n, mean = rep(0, p), sigma = Sigma)
 ```
+
+### Step 2 — Compute the DFT and smoothed periodogram:
+
+```{r}
+# compute DFT at a specific Fourier frequency j
+m <- floor(sqrt(n))   # bandwidth
+j <- 1                # frequency index
+d_j <- dft.X(X_t, j, m)
+
+# estimate spectral density via smoothed periodogram
+f_j_hat <- t(d_j) %*% Conj(d_j) / (2*m + 1)
+```
+
+### Step 3 — Estimate sparse spectral precision matrix via \texttt{cglasso}:
+
+```{r}
+fit_sim <- cglasso(S = f_j_hat, type = "I", nobs = n)
+```
+
+### Step 4 — Visualize the estimated spectral precision matrix:
+
+```{r}
+plot(fit_sim$Theta_list, 
+     index = fit_sim$min_index, 
+     type = "mod", 
+     label = TRUE)
+```
+
+### Step 5 — Compare to the true precision matrix:
+
+```{r}
+# selected lambda
+fit_sim$lambda_grid[fit_sim$min_index]
+
+# number of nonzero off-diagonal entries
+Theta_est <- fit_sim$Theta_list[[fit_sim$min_index]]
+sum(Mod(Theta_est[row(Theta_est) != col(Theta_est)]) > 1e-6)
+```
+
+
 
