+++ vignettes/cxreg_refs.bib
@@ -1,54 +1,0 @@
-@article{friedman2021package,
-  title={Package ‘glmnet’},
-  author={Friedman, Jerome and Hastie, Trevor and Tibshirani, Rob and Narasimhan, Balasubramanian and Tay, Kenneth and Simon, Noah and Qian, Junyang},
-  journal={CRAN R Repositary},
-  volume={595},
-  year={2021}
-}
-
-@article{friedman2010regularization,
-  title={Regularization paths for generalized linear models via coordinate descent},
-  author={Friedman, Jerome and Hastie, Trevor and Tibshirani, Rob},
-  journal={Journal of Statistical Software},
-  volume={33},
-  number={1},
-  pages={1},
-  year={2010},
-  publisher={NIH Public Access}
-}
-
-@article{deb2024regularized,
-  title={Regularized estimation of sparse spectral precision matrices},
-  author={Deb, Navonil and Kuceyeski, Amy and Basu, Sumanta},
-  journal={arXiv preprint arXiv:2401.11128},
-  year={2024}
-}
-
-@article{friedman2008sparse,
-  title={Sparse inverse covariance estimation with the graphical lasso},
-  author={Friedman, Jerome and Hastie, Trevor and Tibshirani, Robert},
-  journal={Biostatistics},
-  volume={9},
-  number={3},
-  pages={432--441},
-  year={2008},
-  publisher={Oxford University Press}
-}
-
-@book{Whittle1951,
-  author    = {Peter Whittle},
-  title     = {Hypothesis Testing in Time Series Analysis},
-  year      = {1951},
-  publisher = {Almqvist \& Wiksells},
-  address   = {Uppsala, Sweden},
-  note      = {Ph.D. thesis, Uppsala University}
-}
-
-@incollection{jankova2018inference,
-  title={Inference in high-dimensional graphical models},
-  author={Jankov{\'a}, Jana and van de Geer, Sara},
-  booktitle={Handbook of graphical models},
-  pages={325--350},
-  year={2018},
-  publisher={CRC Press}
-}
