+++ R/classo.control.R
@@ -1,4 +1,16 @@
-#' internal classo parameters
+# ----------------------------------------------------------------- #
+# Environment-backed settings store so that classo.control(itrace=1)
+# actually persists across calls within the same R session.
+# ----------------------------------------------------------------- #
+#' @noRd
+.classo.env <- new.env(parent = emptyenv())
+.classo.env$settings <- list(
+  fdev   = 1e-05, devmax = 0.999,  eps    = 1e-06, big    = 9.9e+35,
+  mnlam  = 5,     pmin   = 1e-09,  exmx   = 250,   prec   = 1e-10,
+  mxit   = 100,   itrace = 0,      epsnr  = 1e-06, mxitnr = 25
+)
+
+#' Internal classo parameters
 #'
 #' View and/or change the factory default parameters in classo
 #'
@@ -28,85 +40,54 @@
 #' and \code{cv.classo}. factory default = 0
 #' @param epsnr convergence threshold for \code{classo.fit}. factory default =
 #' 1.0e-6
-#' @param mxitnr maximum iterations for the IRLS loop in \code{classo.fit}. factory
-#' default = 25
+#' @param mxitnr maximum iterations for the IRLS loop in \code{classo.fit}.
+#' factory default = 25
 #' @param factory If \code{TRUE}, reset all the parameters to the factory
 #' default; default is \code{FALSE}
 #' @return A list with named elements as in the argument list
 #' @seealso \code{classo}
 #' @keywords models regression
 #' @examples
-#'
-#' classo.control(fdev = 0)  #continue along path even though not much changes
-#' classo.control()  # view current settings
-#' classo.control(factory = TRUE)  # reset all the parameters to their default
-#'
+#' classo.control(fdev = 0)  # continue along path even though not much changes
+#' classo.control()           # view current settings
+#' classo.control(factory = TRUE)  # reset all parameters to their default
 #' @export classo.control
-classo.control <- function (fdev = 1e-05, devmax = 0.999, eps = 1e-06, big = 9.9e+35,
-                            mnlam = 5, pmin = 1e-09, exmx = 250, prec = 1e-10, mxit = 100,
-                            itrace = 0, epsnr = 1e-06, mxitnr = 25,
-                            factory = TRUE
-                            ){
-
-  # ------------------------------------------------ #
-  list(fdev = 1e-05, devmax = 0.999, eps = 1e-06, big = 9.9e+35,
-         mnlam = 5, pmin = 1e-09, exmx = 250, prec = 1e-10, mxit = 100,
-         itrace = 0, epsnr = 1e-06, mxitnr = 25)
-
-  # Change factory=TRUE for now!
-  # Options chosen when FALSE are called by RccpExports.R!
-
-  # inquiry =! nargs()
-  # if (factory){
-  #   invisible(classo.control(fdev = 1e-05, devmax = 0.999,
-  #                            eps = 1e-06, big = 9.9e+35, mnlam = 5, pmin = 1e-09,
-  #                            exmx = 250, prec = 1e-10, mxit = 100, itrace = 0,
-  #                            epsnr = 1e-06, mxitnr = 25))
-  # }else {
-  #   if (!missing(fdev)){
-  #     chg_fract_dev(as.double(fdev))
-  #   }
-  #   if (!missing(devmax)){
-  #     chg_dev_max(as.double(devmax))
-  #   }
-  #   if (!missing(eps)){
-  #     chg_min_flmin(as.double(eps))
-  #   }
[…]
