+++ R/var.cov.R
@@ -1,0 +1,241 @@
+#' Variance-covariance estimation for the debiased spectral precision matrix
+#'
+#' Estimates the asymptotic variance and pseudovariance of the debiased
+#' (desparsified) \code{cglasso} estimator \eqn{\tilde{\Theta}} at a target
+#' Fourier frequency \eqn{\omega_j}, then decomposes them into the
+#' variance of the real part, variance of the imaginary part, and their
+#' cross-covariance. Two estimators are available: a plug-in estimator
+#' based on the theoretical sandwich formula, and a heteroscedasticity- and
+#' autocorrelation-consistent (HAC) estimator.
+#'
+#' For a \eqn{p \times p} complex-valued estimator \eqn{\tilde{\Theta}},
+#' the asymptotic distribution involves both the variance matrix
+#' \eqn{V^{(ab)} = \mathrm{Var}(\tilde{\Theta}_{ab})} and the
+#' pseudovariance matrix \eqn{PV^{(ab)} = \mathrm{PVar}(\tilde{\Theta}_{ab})}.
+#' From these, the variances and covariance of the real and imaginary parts are
+#' recovered via
+#' \deqn{\mathrm{Var}(\mathrm{Re}\,\tilde{\Theta}_{ab}) = \frac{1}{2}\mathrm{Re}(V + PV),}
+#' \deqn{\mathrm{Var}(\mathrm{Im}\,\tilde{\Theta}_{ab}) = \frac{1}{2}\mathrm{Re}(V - PV),}
+#' \deqn{\mathrm{Cov}(\mathrm{Re}\,\tilde{\Theta}_{ab},\,\mathrm{Im}\,\tilde{\Theta}_{ab})
+#'   = \frac{1}{2}\mathrm{Im}(PV).}
+#'
+#' @param Theta A \eqn{nvar \times nvar} complex-valued matrix. The debiased spectral
+#'   precision matrix estimate \eqn{\tilde{\Theta}} at frequency \eqn{\omega_j},
+#'   typically the \code{Theta_tilde} component returned by \code{\link{decglasso}}.
+#' @param X A numeric matrix of size \eqn{nobs \times nvar} (time points by
+#'   variables). The DFT is computed internally via \code{\link{dft.all}}.
+#' @param j An integer frequency index (1-based) identifying the target Fourier
+#'   frequency \eqn{\omega_j = 2\pi j / nobs}.
+#' @param m A positive integer. The half-bandwidth used in the smoothed
+#'   periodogram estimator; must satisfy \eqn{2m + 1 < nobs}.
+#' @param type Character string selecting the variance estimator. Either
+#'   \code{"plug-in"} (default) for the sandwich plug-in estimator, or
+#'   \code{"HAC"} for the heteroscedasticity- and autocorrelation-consistent
+#'   estimator.
+#'
+#' @return A list of class \code{"varcov"} with the following components:
+#' \describe{
+#'   \item{\code{Vre}}{A \eqn{nvar \times nvar} real matrix. Estimated variance of
+#'     the real part of \eqn{\tilde{\Theta}}.}
+#'   \item{\code{Vim}}{A \eqn{nvar \times nvar} real matrix. Estimated variance of
+#'     the imaginary part of \eqn{\tilde{\Theta}}.}
[…]

[…]
